feat(marketing): refactor Code.org footer to use FooterMui
