+++ frontend/apps/marketing/src/components/footer/common/types.ts
@@ -1,0 +1,3 @@
+export interface GlobalFooterProps {
+  locale: string;
+}
