+++ frontend/apps/marketing/src/components/footer/corporateSite/index.ts
@@ -1,0 +1,1 @@
+export {default} from './FooterCorporateSite';
