+++ frontend/apps/marketing/src/components/footer/csForAll/config.tsx
@@ -2,10 +2,10 @@ import Linkedin from '@mui/icons-material/LinkedIn';
 import XIcon from '@mui/icons-material/X';
 
 // Copyright text
-export const COPYRIGHT_TEXT = 'All rights reserved';
+export const COPYRIGHT_TEXT_CSFORALL = 'All rights reserved';
 
-// Main site links
-export const SITE_LINKS = [
+// Site links
+export const SITE_LINKS_CSFORALL = [
   {key: 'initiatives', label: 'Initiatives', href: '/initiatives'},
   {key: 'take-action', label: 'Take Action', href: '/take-action'},
   {key: 'hour-of-ai', label: 'Hour of AI', href: '/hour-of-ai'},
@@ -32,7 +32,7 @@ export const SITE_LINKS = [
 ];
 
 // Social media links
-export const SOCIAL_LINKS = [
+export const SOCIAL_LINKS_CSFORALL = [
   {
     key: 'x-twitter',
     label: 'X',
