+++ frontend/apps/marketing/src/components/footer/csForAll/index.ts
@@ -1,0 +1,1 @@
+export {default} from './FooterCSforAll';
