+++ frontend/apps/marketing/src/components/footer/footerMui/onetrust.scss
