--- frontend/apps/marketing/src/components/footer/index.ts
@@ -1,2 +1,0 @@
-// Export the Footer component for use on frontend/apps/marketing/src/app/[brand]/layout.tsx
-export {default, defaultProps} from './Footer';
