--- frontend/packages/component-library/src/cms/footer/README.md
@@ -1,17 +1,0 @@
-# `component-library/cms/footer`
-
-## Consuming This Component
-
-This package exports one styled React component: [Footer](Footer.tsx).
-You can import it like this:
-
-```javascript
-import Footer from '@code-dot-org/component-library/cms/footer';
-```
-
-_Note:_ This component is used primarily for the Contentful CMS.
-
-For guidelines on how to use this component and the features it
-offers, [visit Storybook](https://code-dot-org.github.io/code-dot-org/component-library-storybook/?path=/docs/cms-footer--docs).
-Or run storybook locally and go
-to [CMS / Footer](http://localhost:6006/?path=/docs/cms-footer--docs).
