--- frontend/packages/component-library/src/cms/footer/footer.module.scss
@@ -1,215 +1,0 @@
-@use '@code-dot-org/component-library-styles/mixins.scss' as mixins;
-@use '@code-dot-org/component-library-styles/typography.module.scss' as
-  typography;
-@use '@code-dot-org/component-library-styles/variables.scss' as variables;
-
-// Footer common styles
-.footer {
-  background: var(--background-neutral-primary-inverse);
-  padding: 1.5rem 1.5rem 1.88rem;
-  z-index: variables.$zindex-fixed;
-
-  .footerContent {
-    display: grid;
-    grid-template-columns: 1fr auto;
-    grid-column-gap: 0.5rem;
-    grid-row-gap: 0.62rem;
-    max-width: 960px;
-    margin: 0 auto;
-
-    .footerLinkListToggle {
-      display: none;
-    }
-
-    .footerLinkList {
-      display: flex;
-      flex-wrap: wrap;
-      list-style: none;
-      margin: 0;
-      padding: 0;
-
-      li {
-        &:not(:last-child)::after {
-          content: '';
-          border: 0.5px solid var(--neutral-white-alpha-20);
-          margin: 0.88rem;
-          font-size: 0.65rem;
-          vertical-align: text-bottom;
-        }
-
-        a {
-          color: var(--text-neutral-inverse);
-          text-decoration: none;
-
-          &:hover {
-            text-decoration: underline;
-          }
-        }
-      }
-    }
-
-    .footerLinkList li,
-    .footerLinkListToggle {
-      @include typography.body-three;
-      @include typography.strong;
-      color: var(--text-neutral-inverse);
-      margin: 0;
-    }
-
-    .footerLanguageDropdown {
-      grid-row: span 3;
-      display: inline-flex;
-    }
-
-    .footerSocialList {
-      display: flex;
-      flex-wrap: wrap;
-      list-style: none;
-      column-gap: 1.25rem;
-      padding: 0;
-      margin: 0;
-
-      li {
-        margin: 0;
-
-        i {
-          @include mixins.icon-dimensions(var(--font-size-body-lg));
-          color: var(--text-neutral-inverse);
-          text-align: center;
-        }
-      }
-    }
-
-    .footerCopyright {
-      @include typography.body-four;
-      color: var(--text-neutral-inverse);
-      white-space: pre-wrap;
-      opacity: 0.8;
-      margin: 0;
-    }
-
-    .footerImageLinkList {
-      display: flex;
-      flex-wrap: wrap;
-      list-style: none;
-      padding: 0;
-      margin: 0.63rem 0 0;
-      gap: 0.5rem;
-
-      li {
-        max-width: 11rem;
-        height: 2.4253rem;
-
-        .footerImageLinkListImg {
-          display: contents;
-
-          img {
-            object-fit: contain;
-          }
-        }
-      }
-    }
-  }
-}
-
-// Footer tablet styles
-@media (max-width: 970px) {
-  .footer {
-    text-align: center;
-
-    .footerContent {
-      grid-template-columns: 1fr;
-      grid-row-gap: 1.12rem;
-      justify-items: center;
-
-      .footerLinkList {
-        justify-content: center;
-      }
-
-      .footerSocialList {
-        grid-row: auto;
-      }
-
-      .footerImageLinkList {
-        margin-top: 0.13rem;
-        justify-content: center;
-      }
-    }
-  }
-}
-
-// Footer mobile styles
-@media (max-width: 600px) {
-  .footer {
-    position: relative;
-    padding-top: 0;
-
-    .footerContent {
-      .footerLinkListToggle {
-        display: flex;
-        justify-content: center;
-        column-gap: 0.38rem;
-        cursor: pointer;
-        width: 100%;
-        margin: 0;
-
-        input[type='checkbox'] {
-          display: none;
-        }
-
-        i {
-          @include mixins.icon-dimensions(var(--font-size-body-sm));
-          @include mixins.transition-all;
-          text-align: center;
-          line-height: 1.54;
-          transform: scaleY(1);
-        }
-
-        & + .footerLinkList {
-          background: var(--background-neutral-primary-inverse);
-          position: absolute;
-          left: 0;
-          bottom: 100%;
-          width: 100%;
-          flex-direction: column;
-          flex-wrap: nowrap;
-          overflow: auto;
-          transition: max-height 0.5s ease-out;
-
-          li {
-            margin: 0 1.5rem;
-
-            &::after {
-              display: none;
-            }
-
-            a {
-              display: block;
-            }
-          }
-        }
-
-        &,
-        & + .footerLinkList li {
-          padding: 0.5rem 0;
-          border-bottom: 1px solid var(--neutral-white-alpha-20);
-        }
-
-        &:has(input:not(:checked)) + .footerLinkList,
-        &:has(input:checked:active) + .footerLinkList {
-          max-height: 0;
-        }
-
-        &:has(input:checked) {
-          i {
-            transform: scaleY(-1);
-          }
-
-          & + .footerLinkList {
-            max-height: 200vh;
-          }
-        }
-      }
-    }
-  }
-}
