--- frontend/packages/component-library/src/cms/footer/index.ts
@@ -1,4 +1,0 @@
-import './index.css';
-
-export type {FooterProps, SiteLink, SocialLink, ImageLink} from './Footer';
-export {default as default} from './Footer';
